Moves darktheme into src folder
