+++ src/Theme.cpp
@@ -17,7 +17,7 @@ AbstractTheme::~AbstractTheme()
 
 // StandardTheme
 
-#include "../darktheme/src/StandardStyle.h"
+#include "../darktheme/StandardStyle.h"
 
 StandardTheme::StandardTheme()
 {
@@ -67,7 +67,7 @@ ThemePtr createStandardTheme()
 
 //#ifdef USE_DARK_THEME
 
-#include "../darktheme/src/DarkStyle.h"
+#include "darktheme/DarkStyle.h"
 
 // DarkTheme
