+++ src/darktheme/StandardStyle.cpp
